requirement.txt
streamlit
pandas
altair
numpy
sklearn.model_selection
sklearn.preprocessing
sklearn.ensemble
imblearn.under_sampling
sklearn.metrics